pyproject.toml
[tool.poetry]
name = "simple-grpc-py"
version = "0.1.0"
description = ""
authors = ["rage puppel"]
readme = "README.md"
packages = [{include = "simple_grpc_py"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
protobuf = "^4.24.4"
loguru = "^0.7.2"
protoletariat = "^3.2.19"
